refactor: move middleware and auth to api/ folder
