+++ gpt2giga/app/factory.py
@@ -4,14 +4,14 @@ from fastapi import Depends, FastAPI
 from starlette.middleware.cors import CORSMiddleware
 from starlette.responses import RedirectResponse
 
+from gpt2giga.api.dependencies.auth import verify_api_key, verify_api_key_gemini
+from gpt2giga.api.middleware.pass_token import PassTokenMiddleware
+from gpt2giga.api.middleware.path_normalizer import PathNormalizationMiddleware
+from gpt2giga.api.middleware.request_validation import RequestValidationMiddleware
+from gpt2giga.api.middleware.rquid_context import RquidMiddleware
 from gpt2giga.app.cli import load_config
 from gpt2giga.app.lifespan import lifespan
-from gpt2giga.auth import verify_api_key, verify_api_key_gemini
 from gpt2giga.core.app_meta import get_app_version
-from gpt2giga.middlewares.pass_token import PassTokenMiddleware
-from gpt2giga.middlewares.path_normalizer import PathNormalizationMiddleware
-from gpt2giga.middlewares.request_validation import RequestValidationMiddleware
-from gpt2giga.middlewares.rquid_context import RquidMiddleware
 from gpt2giga.protocol.gemini.response import GeminiAPIError, gemini_error_response
 from gpt2giga.routers.anthropic import router as anthropic_router
 from gpt2giga.routers.gemini import router as gemini_router
